fix: パラメータの設定を別ファイルに分離
